Refactors rake tasks and install generator used to setup and fixes Travis bundle
Changes .travis.yml to use a container ( docker ) to run flack. This was primarily motivated by the bundler issues we had on Travis, where we had errors related to missing gems.
For example, we'd install our dependencies, move to another directory, clone flack, install his dependencies using bundler, and then we'd run make migrate, which would fail on missing dependencies.

Running flack this way also reinforces the idea that flor is a completely decoupled system with which we interact over HTTP.

For our specs, flack remains installed and ran locally, for simplicity's sake.

Adds the steps to go from clone to test ( rspec ) in README.md.